Update dependencies and don't use proc_macro_hack (remove impl crate)
