--- impl/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "vk-shader-macros-impl"
-description = "Implementation detail of vk-shader-macros"
-version = "0.2.7"
-authors = ["Benjamin Saunders <ben.e.saunders@gmail.com>"]
-license = "MIT/Apache-2.0"
-edition = "2018"
-
-[lib]
-proc-macro = true
-
-[features]
-strip = []
-default-optimize-zero = []
-build-from-source = ["shaderc/build-from-source"]
-
-[dependencies]
-syn = { version = "1.0.2", default-features = false, features = [ "parsing", "proc-macro", "derive" ] }
-quote = "1.0.1"
-proc-macro2 = "1.0.1"
-proc-macro-hack = "0.5.4"
-shaderc = "0.7.2"
